+++ index.html
@@ -75,7 +75,7 @@
         </div>
     </div>
 
-    <script src="src/index.js"></script>
+    <script src="src/frontend.js"></script>
     <!--
         <script src="build/bundle.js"></script>
         <script src="../src/init.js"></script>
